+++ inc/form.php
@@ -200,7 +200,7 @@ class form {
 
 function form_process_def(&$def) {
   foreach($def as $k=>$element_def) {
-    if(isset($element_def['count'])&&($element_def['type']!="array")) {
+    if(isset($element_def['count'])&&(!in_array($element_def['type'], array("array", "hash")))) {
       $def[$k]=$element_def['count'];
 
       if(!is_array($def[$k])) {