+++ lang/de.php
@@ -26,3 +26,6 @@ $lang_str['form_element:please_select'] = "-- bitte wählen --";
 $lang_str['form:check_all'] = "alle markieren";
 $lang_str['form:uncheck_all'] = "keine markieren";
 $lang_str['form:duplicate'] = array("Wert %s kommt mehrfach vor.", "Werte %s kommen mehrfach vor.");
+
+$lang_str['form:hash_key_field_name'] = "Schlüssel";
+$lang_str['form:hash_value_field_name'] = "Wert";